pfblockerng: ADR-08 inline the IDN homoglyph analyzer into pfb_unbound.py
The name()-based analyzer is ~90 executable lines, so fold it into pfb_unbound.py
(the production matcher module) and delete the separate pfb_idn_analyzer.py. It now
rides pfb_unbound.py's existing Unbound-chroot copy and FreeBSD-ports pkg-plist entry,
so there is NO second shipped file, NO extra chroot-copy wiring (the pfblockerng.inc /
pfblockerng_install.inc copy/unlink/glob additions are reverted), and NO pkg-plist
change — both deployment concerns from the review evaporate.

Behaviour is unchanged: the test suite imports the analyzer symbols from pfb_unbound
(conftest injects the Unbound runtime globals so it imports off-appliance); 1307 passed,
PHPStan/PHPUnit/php -l/ruff/markdownlint clean. Analyzer docstrings streamlined while
moving. Docs (README/CLAUDE.md/ADR §7) updated for the inlined, no-separate-file reality.
