+++ README.md
@@ -317,8 +317,8 @@ caught. See [ADR-08](.ADRs/ADR_08_Homoglyph_Protection/ADR.md) for the full cont
 the decision oracle + analyzer + matcher-wiring + corpus guards are the
 `tests/test_adr08_*` suite.
 
-The analyzer (`src/usr/local/pkg/pfblockerng/pfb_idn_analyzer.py`) resolves each code
-point's script from the **stdlib `unicodedata.name()`** leading token (`LATIN…`,
+The analyzer (inlined in `src/usr/local/pkg/pfblockerng/pfb_unbound.py`) resolves each
+code point's script from the **stdlib `unicodedata.name()`** leading token (`LATIN…`,
 `CYRILLIC…`, `GREEK…`, `CJK…`→Han, …) — no shipped Unicode table and no runtime
 dependency. The script names are stable across UCD versions for the established
 scripts in scope, so the FP/TP result holds across Python 3.11–3.14 (validated in the
